Move capability related classes to the `capabilities` package
