+++ subprojects/core-api/src/main/java/org/gradle/api/capabilities/CapabilitiesMetadata.java
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.api.artifacts;
+package org.gradle.api.capabilities;
 
 import org.gradle.api.Incubating;
-import org.gradle.api.capabilities.CapabilityDescriptor;
 
 import java.util.List;
