+++ subprojects/core-api/src/main/java/org/gradle/api/capabilities/MutableCapabilitiesMetadata.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.api.artifacts;
+package org.gradle.api.capabilities;
 
 import org.gradle.api.Incubating;
