+++ subprojects/dependency-management/src/main/java/org/gradle/internal/component/model/CapabilitiesRules.java
@@ -17,8 +17,8 @@ package org.gradle.internal.component.model;
 
 import com.google.common.collect.Lists;
 import org.gradle.api.Action;
-import org.gradle.api.artifacts.CapabilitiesMetadata;
-import org.gradle.api.artifacts.MutableCapabilitiesMetadata;
+import org.gradle.api.capabilities.CapabilitiesMetadata;
+import org.gradle.api.capabilities.MutableCapabilitiesMetadata;
 import org.gradle.internal.component.external.model.VariantMetadataRules;
 
 import java.util.List;
